+++ tools/build_variables.bzl
@@ -88,6 +88,7 @@ libtorch_core_sources = [
     "torch/csrc/jit/api/function_impl.cpp",
     "torch/csrc/jit/api/module.cpp",
     "torch/csrc/jit/api/object.cpp",
+    "torch/csrc/jit/backends/backend_detail.cpp",
     "torch/csrc/jit/backends/backend_interface.cpp",
     "torch/csrc/jit/codegen/fuser/codegen.cpp",
     "torch/csrc/jit/codegen/fuser/compiler.cpp",
@@ -421,7 +422,7 @@ libtorch_python_core_sources = [
     "torch/csrc/autograd/python_legacy_variable.cpp",
     "torch/csrc/autograd/python_variable.cpp",
     "torch/csrc/autograd/python_variable_indexing.cpp",
-    "torch/csrc/jit/backends/backend_detail.cpp",
+    "torch/csrc/jit/backends/backend.cpp",
     "torch/csrc/jit/backends/backend_init.cpp",
     "torch/csrc/jit/backends/backend_resolver.cpp",
     "torch/csrc/jit/backends/test_backend.cpp",
